Refactor the orbitdb code to make a postgres implementation easier
