+++ data/orbitdb/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "orbitdb",
+  "name": "blockchain-carbon-accounting-orbitdb",
   "version": "1.0.0",
   "description": "Testing out OrbitDB for use in blockchain-carbon-accounting project",
   "main": "src/dataLoader.ts",
@@ -29,19 +29,20 @@
     "orbit-db-io": "^1.0.2",
     "orbit-db-store": "^4.3.3",
     "uuid": "^8.3.2",
-    "xlsx": "^0.18.3",
+    "xlsx": "^0.18.4",
     "yargs": "^17.3.1"
   },
   "devDependencies": {
     "@types/cli-progress": "^3.9.2",
     "@types/node": "^17.0.21",
     "@types/orbit-db": "github:orbitdb/orbit-db-types",
+    "@types/uuid": "^8.3.4",
     "@types/yargs": "^17.0.9",
-    "@typescript-eslint/eslint-plugin": "^5.14.0",
-    "@typescript-eslint/parser": "^5.14.0",
-    "esbuild": "^0.14.25",
+    "@typescript-eslint/eslint-plugin": "^5.15.0",
+    "@typescript-eslint/parser": "^5.15.0",
+    "esbuild": "^0.14.27",
     "esbuild-runner": "^2.2.1",
-    "eslint": "^8.10.0",
+    "eslint": "^8.11.0",
     "typescript": "^4.6.2"
   }
 }
