+++ data/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "blockchain-carbon-accounting-data",
+  "version": "1.0.0",
+  "description": "Common data module for use in blockchain-carbon-accounting project",
+  "engines": {
+    "node": ">=16"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@types/cli-progress": "^3.9.2",
+    "@types/uuid": "^8.3.4",
+    "@typescript-eslint/eslint-plugin": "^5.15.0",
+    "@typescript-eslint/parser": "^5.15.0",
+    "esbuild": "^0.14.27",
+    "esbuild-runner": "^2.2.1",
+    "eslint": "^8.11.0",
+    "typescript": "^4.6.2"
+  },
+  "dependencies": {
+    "cli-progress": "^3.10.0",
+    "uuid": "^8.3.2",
+    "xlsx": "^0.18.4"
+  }
+}
